MealPlanComponent list now leverages RxJS
Created a new MealPlanListDatasource service. This is responsible for
handling the data for the meal plan list.

It leverages a BehaviorSubject for its data handling. It's still a bit
imperative it in its design, but that does make it easy to follow. Going
full declarative is possible by leveraging more complicated operators
from RxJS, but it doesn't feel worth it.

This could probably still be improved by storing the data as a map to
allow more efficient updates, but this should be fine for now.

NOTE: this is completely untested.